requirements.txt
python-dotenv
pandas
numpy
requests
torch
